kids-story-generator/requirements.txt
openai~=1.70.0
python-dotenv~=1.1.0
pillow~=11.2.1
requests~=2.32.3
streamlit~=1.37.0
fpdf~=1.7.2
